==> 03-compile-pheno/00-check-field.ipynb <==
from itertools import compress
import pandas as pd
import itertools

filepath = '/u/home/z/ziqixu09/project-pasaniuc/phenotype/prepare-pheno-fields.R'
field_dict = {}

# create a dict to store all the fields we already have
with open(filepath) as fp:
    line = fp.readline()
    i = 0
    while line:
        if "-0.0" in line:
            li = line.split('"')
            split_li = ['-0.0' in ele for ele in li]
            name = list(compress(li, split_li))
            field_dict[i] = name
        if "-2.0" in line:
            li2 = line.split('"')
            split_li2 = ['-2.0' in ele for ele in li2]
            name2 = list(compress(li2, split_li2))
            field_dict[i] = name2
        line = fp.readline()
        i += 1

# extract all values from the dict and remove duplicates
dict_val = list(field_dict.values())
chain = itertools.chain(*dict_val)
val_li = list(chain)
val_li = list(set(val_li))

# read the head of two phenotype files
hd1 = pd.read_csv("/u/home/z/ziqixu09/project-pasaniuc/phenotype/head1.csv")
hd1 = hd1.columns.values.tolist()
hd2 = pd.read_csv("/u/home/z/ziqixu09/project-pasaniuc/phenotype/head2.csv")
hd2 = hd2.columns.values.tolist()

# conbine the head of two files, check if all fields are contained
hd = hd1 + hd2
hd_contain1 = ['-0.0' in ele for ele in hd]
hd_field1 = list(compress(hd, hd_contain1))
hd_contain2 = ['-2.0' in ele for ele in hd]
hd_field2 = list(compress(hd, hd_contain2))
hd_pheno = hd_field1 + hd_field2
intersection_set = set.intersection(set(val_li), set(hd_pheno))
intersection_list = list(intersection_set)

# check separately 
# if the first phenotype file contains all the fields
hd1_contain1 = ['-0.0' in ele for ele in hd1]
hd1_field1 = list(compress(hd1, hd1_contain1))
hd1_contain2 = ['-2.0' in ele for ele in hd1]
hd1_field2 = list(compress(hd1, hd1_contain2))
hd1_pheno = hd1_field1 + hd1_field2
intersection_set1 = set.intersection(set(val_li), set(hd1_pheno))
intersection_list1 = list(intersection_set1)

# if the second phenotype file contains the fields
hd2_contain1 = ['-0.0' in ele for ele in hd2]
hd2_field1 = list(compress(hd2, hd2_contain1))
hd2_contain2 = ['-2.0' in ele for ele in hd2]
hd2_field2 = list(compress(hd2, hd2_contain2))
hd2_pheno = hd2_field1 + hd2_field2
intersection_set2 = set.intersection(set(val_li), set(hd2_pheno))
intersection_list2 = list(intersection_set2)

# check if the hit fields in second phenotype file contains the hit fields in first phenotype file
dup_set = set.intersection(intersection_set1, intersection_set2)
dup_set == intersection_set1

# check if the original fields are all contained 
set(val_li) == intersection_set2

# add for a list of logMAR traits
li1 = [f"5078-0.{i}" for i in range(0, 16)]
li2 = [f"5078-1.{i}" for i in range(0, 16)]
li3 = [f"5079-0.{i}" for i in range(0, 16)]
li4= [f"5079-1.{i}" for i in range(0, 16)]

# output the hit fields
hits_li = intersection_list2 + li1+ li2+ li3 +li4
df = pd.DataFrame(hits_li)
df.to_csv("hits.csv",index=False)
