==> 03-compile-pheno/00-check-phecode.ipynb <==
from itertools import compress
import pandas as pd
import itertools

filepath = '/u/home/z/ziqixu09/project-pasaniuc/phenotype/prepare-phecodes.R'
field_dict = {}

# create a dict to store all the fields we already have
with open(filepath) as fp:
    line = fp.readline()
    i = 0
    while line:
        if ".0" in line:
            li = line.split('"')
            split_li = ['.0' in ele for ele in li]
            name = list(compress(li, split_li))
            field_dict[i] = name
        if "-1." in line:
            li2 = line.split('"')
            split_li2 = ['-1.' in ele for ele in li2]
            name2 = list(compress(li2, split_li2))
            field_dict[i] = name2
        if "-0." in line:
            li2 = line.split('"')
            split_li2 = ['-0.' in ele for ele in li2]
            name2 = list(compress(li2, split_li2))
            field_dict[i] = name2
        line = fp.readline()
        i += 1

field_dict

field_dict.pop(16)
field_dict.pop(17)

field1 = [f"40001-{i}.0" for i in range(2)]
field2 = [f"40002-0.{i}" for i in range(14)]
field3 = [f"40002-1.{i}" for i in range(14)]
field4 = [f"40006-{i}.0" for i in range(17)]
field5 = [f"41270-0.{i}" for i in range(213)]
field6 = [f"41201-0.{i}" for i in range(22)]
field7 = [f"40013-{i}.0" for i in range(15)]
field8 = [f"41271-0.{i}" for i in range(47)]
field9 = ["22001-0.0"]
field10 = [f"20544-0.{i}" for i in range(17)]
val_li = field1 + field2 + field3 + field4 + field5 + field6 + field7 + field8 + field9 + field10

len(list(set(val_li)))

# read the head of two phenotype files
hd1 = pd.read_csv("/u/home/z/ziqixu09/project-pasaniuc/phenotype/head1.csv")
hd1 = hd1.columns.values.tolist()
hd2 = pd.read_csv("/u/home/z/ziqixu09/project-pasaniuc/phenotype/head2.csv")
hd2 = hd2.columns.values.tolist()

len(hd1)

check1 = set.intersection(set(val_li), set(hd2))
check2 = set.intersection(set(val_li), set(hd1))
hits = set(sorted(list(set(check1)) + list(set(check2))))

len(hits)

# check if every hit is contained in head2
dup_set = set.intersection(check1, check2)
dup_set == check2

df = pd.DataFrame(hits)
df.to_csv("phecode_hits.csv",index=False)
