==> 03-compile-pheno/01-extract-phecode.ipynb <==
import numpy as np
import pandas as pd

# manage space 
from dask.distributed import Client
import dask
client = Client(processes=False, threads_per_worker=1, n_workers=1, memory_limit='100GB')

# read the hit list of all fields
hit = pd.read_csv("phecode_hits.csv")
hit_li = hit["0"].values
hit_li = np.concatenate((['eid'], hit_li))

# find the sex column and delete it
sex = np.where(hit_li == "22001-0.0")
hit_li2 = hit_li[sex[0]]
hit_li1 = np.delete(hit_li, sex)

# check first phenotype field
path1 = "/u/project/sgss/UKBB/PRS-RESEARCH/03-compile-pheno/raw/ukb39967.enc_ukb.converted2.csv"
file1 = pd.read_csv(path1, usecols = hit_li1, dtype = str, encoding = 'unicode_escape')
file1 = file1.set_index("eid")
file1.index.name = "eid"
file1.index = file1.index.astype('int64')

# check second phenotype field
path2 = "/u/project/sgss/UKBB/PRS-RESEARCH/03-compile-pheno/raw/ukb21970.csv"
col_age = "21003-0.0"
col_sex = "31-0.0"
col_pcs = [f"22009-0.{i}" for i in range(1, 41)]
col_report_ethnic = "21000-0.0"
col_id = "eid"
col_dep = "189-0.0"

cols = [col_id, col_age, col_sex, col_report_ethnic, col_dep] + col_pcs
col_names = ["ID", "AGE", "SEX", "SIRE_ID", "DEPRIVATION_INDEX"] + [f"PC{i + 1}" for i in range(len(col_pcs))]

file2 = pd.read_csv(path2, usecols = cols, encoding = 'unicode_escape')[cols]
file2.columns = col_names
file2 = file2.set_index("ID")
file2.index.name = "eid"

# drop the rows with na values in PC
file2 = file2.dropna(axis = 0, subset = ["PC1"])

# extract the common individual ids of these two files
# ends up 488266 indivs in total
idx = file1.index.intersection(file2.index)
file1 = file1.loc[idx]
file2 = file2.loc[idx]

# append the sex column to phecode file
file1["22001-0.0"] = file2["SEX"]

file1.to_csv("phecode.csv", na_rep='NA')
file2.to_csv("covar.csv", na_rep='NA')
