==> 03-compile-pheno/02-format.ipynb <==
%load_ext lab_black

import numpy as np
import pandas as pd
import os
import pyreadr
from glob import glob
import matplotlib.pyplot as plt

PHENO_DIR = "/u/project/pasaniuc/ziqixu09/phenotype/"

# extract covariates
phecode_eid = pd.read_csv(os.path.join(PHENO_DIR, "phecode.csv"), usecols=["eid"])[
    "eid"
].values.astype(str)

pheno_eid = pd.read_csv(os.path.join(PHENO_DIR, "pheno.csv"), usecols=["eid"])[
    "eid"
].values.astype(str)

assert np.all(phecode_eid == pheno_eid)

df_covar = pd.read_csv(os.path.join(PHENO_DIR, "covar.csv"), index_col=0)[
    ["AGE", "SEX", "DEPRIVATION_INDEX"] + [f"PC{i}" for i in range(1, 17)]
]
df_covar.index = df_covar.index.astype(str)
df_covar = pd.merge(
    pd.DataFrame(
        {"FID": df_covar.index, "IID": df_covar.index},
        index=df_covar.index,
    ),
    df_covar,
    left_index=True,
    right_index=True,
)
df_covar.to_csv(
    "out/covar.tsv", sep="\t", index=False, float_format="%.6g", na_rep="NA"
)
assert np.all(df_covar.index == pheno_eid)

df_pheno = df_covar[["FID", "IID"]].copy()
df_pheno["PHENO"] = np.nan

print("Proportion of non-NaNs: ")
for trait_type in ["quant", "binary"]:
    trait_dir = f"data/ukbb-{trait_type}-pheno/"
    for trait_f in glob(f"{trait_dir}/*.rds"):
        trait = trait_f.split("/")[-1][:-4]
        trait_values = pyreadr.read_r(trait_f)[None].values
        df_pheno["PHENO"] = trait_values.astype(float)
        print(f"{trait}: {1 - df_pheno['PHENO'].isna().mean():.2g}")
        df_pheno.to_csv(
            f"out/{trait}.tsv",
            sep="\t",
            index=False,
            float_format="%.6g",
            na_rep="NA",
        )

# check the consistency with old version of covar
df_covar_old = pd.read_csv(
    "/u/home/k/kangchen/admix-prs-uncertainty/experiments/00-compile-data/out/covar.tsv",
    sep="\t",
    index_col=0,
)
df_covar_old.index = df_covar_old.index.astype(str)
assert np.all(df_covar_old.index == df_covar.index)
for col in set(df_covar_old.columns) & set(df_covar.columns):
    assert np.allclose(df_covar_old[col], df_covar[col], equal_nan=True)

# check the consistency with old version of pheno

trait_list = [t.split("/")[-1][:-4] for t in glob("out/OLD/*.tsv")]
trait_list = set(trait_list) - set(["covar"])
for trait in trait_list:
    df1 = pd.read_csv(f"out/OLD/{trait}.tsv", sep="\t")
    df2 = pd.read_csv(f"out/{trait}.tsv", sep="\t")
    assert df1.equals(df2)
